% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{applyStaticGating}
\alias{applyStaticGating}
\title{applyStaticGating}
\usage{
applyStaticGating(wsp_path, fcs_path)
}
\arguments{
\item{wsp_path}{Path of the flowjo workspace}

\item{fcs_path}{Path of the fcs file}
}
\value{
A matrix of TRUE/FALSE values, with one row for every cell and one
        column for every gate.
}
\description{
Parse a flowjo workspace and apply the first gating on a different file
}
