% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytofNorm.R
\name{CytofNorm.train}
\alias{CytofNorm.train}
\title{CytofNorm.train}
\usage{
CytofNorm.train(files, labels, channels, transformList,
  outputDir = "./tmp", clean = TRUE, plot = FALSE, verbose = FALSE,
  seed = NULL, FlowSOM.params = NULL,
  normMethod.train = QuantileNorm.train, normParams = list(nQ = 21,
  spar = 0.5), ...)
}
\arguments{
\item{files}{Full paths of to the fcs files of the control samples.}

\item{labels}{A label for every file, indicating to which batch it
belongs, e.g. the plate ID.}

\item{channels}{Column names of the channels that need to be normalized}

\item{transformList}{Transformation list to pass to the flowCore
\code{transform} function}

\item{outputDir}{Directory to put the temporary files in. Default = "./tmp"}

\item{clean}{Whether to remove the temporary files again at the end.
Default = TRUE.}

\item{plot}{If TRUE, plots are saved to the current dir.
Default = FALSE.}

\item{verbose}{If TRUE, progress updates are printed. Default = FALSE.}

\item{seed}{Set a seed for reproducable results.}

\item{FlowSOM.params}{Extra parameters to be passed to the FlowSOM
function, such as the number of cells
(\code{nCells}), the \code{xdim} and \code{y dim} or
the number of clusters (\code{nClus})}

\item{normMethod.train}{Normalization method to use for each cluster.
Default = \code{\link{QuantileNorm.train}}}

\item{normParams}{Parameters to pass to the normalization method. Default,
assuming \code{\link{QuantileNorm.train}}:
list(nQ = 21, spar = 0.5)). nQ is the number of quantiles
to use, spar is the spline smoothing paramter, typically
between (0,1]. See \code{smooth.spline} from the
\code{stats} package.}
}
\value{
A list containing two elements: the FlowSOM clustering and a list
        containing all the splines per cluster. This can be used as input
        for the \code{\link{CytofNorm.normalize}} function.
}
\description{
CytofNorm learns batch effects, using control samples which should be the
the same per batch. The cells are first split into rough groups using the
FlowSOM algorithm, after which splines are trained to map the quantiles for
each control group to an average distribution
Typically, you will use the function \code{\link{CytofNorm.normalize}} after
this function to reverse the batch effects in other files.
}
\details{
Temporary fcs files splitting the data into clusters are written to
\code{outputDir} and will be removed again by default (depending on
\code{clean}).
}
\examples{

}
\seealso{
\code{\link{QuantileNorm.train}}, \code{\link{CytofNorm.normalize}},
           \code{\link{prepareFlowSOM}}
}
