% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{cytofTransform.reverse}
\alias{cytofTransform.reverse}
\title{cytofTransform.reverse}
\usage{
cytofTransform.reverse(x)
}
\arguments{
\item{x}{Value to tranform
Function to reverse arcsinh transformation with cofactor 5}
}
\description{
cytofTransform.reverse
}
