% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoNorm.R
\name{CytoNorm.normalize}
\alias{CytoNorm.normalize}
\title{normalizeClustered}
\usage{
CytoNorm.normalize(model, files, labels, transformList,
  transformList.reverse, outputDir = ".", prefix = "Norm_",
  clean = TRUE, verbose = FALSE,
  normMethod.normalize = QuantileNorm.normalize)
}
\arguments{
\item{model}{Model of the batch effercts, as computed by
\code{\link{CytoNorm.train}}}

\item{files}{Full paths of to the fcs files of the samples.}

\item{labels}{A label for every file, indicating to which batch it
belongs, e.g. the plate ID.}

\item{transformList}{Transformation list to pass to the flowCore
\code{transform} function}

\item{transformList.reverse}{Transformation list with the reverse functions,
so the normalized files can be saved in the untransformed
space}

\item{outputDir}{Directory to put the temporary files in. Default = "."}

\item{prefix}{Prefix to put in front of the normalized file names.
Default = "Norm_"}

\item{clean}{If FALSE, temporary files describing the FlowSOM clusters
seperately are not removed at the end. Default = TRUE.}

\item{verbose}{If TRUE, progress updates are printed. Default = FALSE.}

\item{normMethod.normalize}{Normalization method to use.}
}
\value{
Nothing is returned, but the new FCS files are written to the output
        directory
}
\description{
Normalize data, given the batch effects learned from control samples per
cell type/cluster (output from \code{\link{CytoNorm.train}}). New fcs files
are written to the given output directory.
}
\examples{

}
\seealso{
\code{\link{CytoNorm.train}}
}
