% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoNorm2-package.R
\docType{package}
\name{CytoNorm2-package}
\alias{CytoNorm2}
\alias{CytoNorm2-package}
\title{CytoNorm2: Normalisation of cytometry data measured across multiple batches}
\description{
To learn more about using the CytoNorm package, please start by browsing the
vignettes:
}
\author{
\strong{Maintainer}: Sofie Van Gassen \email{sofie.vangassen@ugent.be}

Authors:
\itemize{
  \item Katrien Quintelier \email{KatrienL.Quintelier@UGent.be}
}

}
\keyword{internal}
