% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoNorm.R
\name{getCytoNormQuantiles}
\alias{getCytoNormQuantiles}
\title{Extract quantiles used by a previous CytoNorm model}
\usage{
getCytoNormQuantiles(model, type = "ref")
}
\arguments{
\item{model}{Model trained with CytoNorm.train}

\item{type}{"ref" for goal quantiles or batch label}
}
\description{
Can be of interest to pass as goal quantiles to a new model.
}
\examples{
dir <- system.file("extdata", package = "CytoNorm")
files <- list.files(dir, pattern = "fcs$")
data <- data.frame(File = files,
                   Path = file.path(dir, files),
                   Type = stringr::str_match(files, "_([12]).fcs")[,2],
                   Batch = stringr::str_match(files, "PTLG[0-9]*")[,1],
                   stringsAsFactors = FALSE)
data$Type <- c("1" = "Train", "2" = "Validation")[data$Type]
train_data <- dplyr::filter(data, Type == "Train")
validation_data <- dplyr::filter(data, Type == "Validation")

ff <- flowCore::read.FCS(data$Path[1])
channels <- grep("Di$", flowCore::colnames(ff), value = TRUE)
transformList <- flowCore::transformList(channels,
                                         cytofTransform)
transformList.reverse <- flowCore::transformList(channels,
                                                 cytofTransform.reverse)

model <- CytoNorm.train(files = train_data$Path,
                        labels = train_data$Batch,
                        channels = channels,
                        transformList = transformList,
                        FlowSOM.params = list(nCells = 10000, #1000000
                                              xdim = 10,
                                              ydim = 10,
                                              nClus = 10,
                                              scale = FALSE),
                        normParams = list(nQ = 99),
                        seed = 1)

quantiles <- getCytoNormQuantiles(model)

}
