% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{getDensities}
\alias{getDensities}
\title{getDensities}
\usage{
getDensities(files, channels, quantileValues = 0.5,
  transformList = NULL, selection = NULL, ...)
}
\arguments{
\item{files}{Full paths of to the fcs files of the samples}

\item{channels}{Names of the channels to compute the densities for}

\item{transformList}{Transformation list to pass to the flowCore
\code{transform} function}

\item{selection}{List with indexation vector for every file.}

\item{...}{Extra parameters to pass to density}
}
\description{
Returns a dataframe with 4 columns, containing for every file and every
channel of interest the x/y values computed by density.
}
\examples{

dir <- system.file("extdata", package = "CytoNorm")
files <- list.files(dir, pattern = "fcs$")

ff <- flowCore::read.FCS(file.path(dir, files[1]))
channels <- grep("Di$", flowCore::colnames(ff), value = TRUE)
transformList <- flowCore::transformList(channels,
                                         cytofTransform)

densities <- getDensities(files = file.path(dir, files),
                          channels = channels,
                          transformList = transformList,
                          quantileValues = c(0.25, 0.5, 0.75))

densities$densities$Marker <- paste0(
  FlowSOM::get_markers(ff, densities$densities$Channel),
  " (", densities$densities$Channel, ")")

ggplot(dplyr::filter(densities$densities,
                     File == file.path(dir, files[1]))) +
  geom_line(aes(x = x, y = y)) +
  facet_wrap(~ Marker, scales = "free") +
  theme_minimal()

ggplot(dplyr::filter(densities$quantiles,
                     Channel == FlowSOM::get_channels(ff, "CD66"))) +
  geom_vline(aes(xintercept = `0.25`), col = "grey") +
  geom_vline(aes(xintercept = `0.5`), col = "grey", lwd = 1) +
  geom_vline(aes(xintercept = `0.75`), col = "grey") +
  geom_line(aes(x = x, y = y),
            data = dplyr::filter(densities$densities,
                     Channel == FlowSOM::get_channels(ff, "CD66"))) +
  facet_wrap(~ File) +
  theme_minimal()

}
