% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{jaccard}
\alias{jaccard}
\title{jaccard}
\usage{
jaccard(a, b)
}
\arguments{
\item{a}{logical vector}

\item{b}{logical vector}
}
\description{
Computes the jaccard index given two logical vectors of equal length:
intersection / union
}
\examples{
a <- c(TRUE, TRUE, FALSE, FALSE, TRUE)
b <- c(TRUE, FALSE, TRUE, FALSE, TRUE)
jaccard(a, b)

}
