% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MinMaxNorm.R
\name{MinMaxNorm.train}
\alias{MinMaxNorm.train}
\title{MinMaxNorm.train}
\usage{
MinMaxNorm.train(files, labels, channels, transformList,
  minmax_q = c(0.001, 0.999), goal = "mean", verbose = FALSE,
  plot = FALSE, plotTitle = "QMinMax")
}
\arguments{
\item{files}{Full paths of to the fcs files of the control samples}

\item{labels}{A label for every file, indicating to which batch it
belongs, e.g. the plate ID.}

\item{channels}{Names of the channels to normalize}

\item{transformList}{Transformation list to pass to the flowCore
\code{transform} function}

\item{minmax_q}{Quantiles to approximate minimum and maximum. Use c(0,1)
for exact minimum and maximum. Default = c(0.001, 0.999)}

\item{goal}{Goal distribution. Default "mean", can also be two numeric
values or one of the batch labels.}

\item{verbose}{Extra output is printed when TRUE. Default FALSE.}

\item{plot}{If TRUE, a plot is generated (using the \code{layout}
function) showing all quantiles. Default = FALSE.}

\item{plotTitle}{Title to use in the plot. Default = "QMinMax".}
}
\value{
A list containing the min max information. This can
        be used as input for the \code{\link{MinMaxNorm.normalize}} function.
}
\description{
Learn the batch effects from control samples.
This function computes min and max to describe the distribution of the data,
and infers spline functions to equalize these distributions over the files.
Typically, you will use the function \code{\link{QuantileNorm.normalize}}
after this function to reverse the batch effects on other files.
}
\examples{

}
\seealso{
\code{\link{MinMaxNorm.normalize}}
}
