% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{named.list}
\alias{named.list}
\title{named.list}
\usage{
named.list(...)
}
\arguments{
\item{...}{variables to add in the list}
}
\description{
Creates a list, in which the names of the items correspond to the names
of the variables.
}
