% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytofNorm.R
\name{prepareFlowSOM}
\alias{prepareFlowSOM}
\title{prepareFlowSOM}
\usage{
prepareFlowSOM(files, transformList, colsToUse, nClus = 30, xdim = 15,
  ydim = 15, nCells = 1e+06, plot = FALSE, seed = NULL,
  FlowSOM.params = NULL)
}
\arguments{
\item{files}{Files to use (with path)}

\item{transformList}{Transformation list to pass to the flowCore \code{transform}
function}

\item{colsToUse}{IDs or column names of the columns to use for clustering}

\item{nClus}{Number of metaclusters FlowSOM should detect. Default = 30.}

\item{nCells}{The total number of cells to use for the FlowSOM clustering.
This number is divided by the number of files to determine
the amount to select from each individual file.
Default = 1 000 000.}

\item{plot}{If TRUE, the FlowSOM tree is plotted. Default = FALSE.}

\item{seed}{If not NULL, set.seed is called with this argument for
reproducable results. Default = NULL.}

\item{gridSize}{Grid dimension to use. In the first clustering step,
dim x dim clusters will be detected. Default = 15.}
}
\value{
FlowSOM object describing the FlowSOM clustering
}
\description{
Aggregate files, transform them and run FlowSOM.
This is used as a first step in the normalization procedure to detect
groups of similar cells. Typically you will not call this function, but use
the wrapper function \code{\link{CytofNorm.train}} instead.
}
\examples{

}
\seealso{
\code{\link{CytofNorm.train}}
}
