% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantileNorm.R
\name{QuantileNorm.normalize}
\alias{QuantileNorm.normalize}
\title{QuantileNorm.normalize}
\usage{
QuantileNorm.normalize(model, files, labels, transformList,
  transformList.reverse, outputDir = ".", prefix = "Norm_",
  removeOriginal = FALSE, verbose = FALSE)
}
\arguments{
\item{model}{Model of the batch effercts, as computed by
\code{\link{QuantileNorm.train}}}

\item{files}{Full paths of to the fcs files of the samples to normalize}

\item{labels}{A label for every file, indicating to which batch it
belongs, e.g. the plate ID.}

\item{transformList}{Transformation list to pass to the flowCore
\code{transform} function}

\item{transformList.reverse}{Transformation list with the reverse functions,
so the normalized files can be saved in the untransformed
space}

\item{outputDir}{Directory to put the temporary files in. Default = "."}

\item{prefix}{Prefix to put in front of the normalized file names.
Default = "Norm_"}

\item{removeOriginal}{Should the original fcs be removed? Default = FALSE.}

\item{verbose}{If TRUE, progress updates are printed. Default = FALSE.}
}
\value{
Nothing is returned, but the new FCS files are written to the output
        directory
}
\description{
Normalize data, given the batch effects learned from control samples per
cell type/cluster (output from \code{\link{QuantileNorm.train}}). New fcs
files are written to the given output directory.
}
\examples{

}
\seealso{
\code{\link{QuantileNorm.train}}
}
