% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantileNorm.R
\name{QuantileNorm.train}
\alias{QuantileNorm.train}
\title{QuantileNorm.train}
\usage{
QuantileNorm.train(files, labels, channels, transformList, nQ = 21,
  spar = 0.5, verbose = FALSE, plot = FALSE,
  plotTitle = "Quantiles")
}
\arguments{
\item{files}{Full paths of to the fcs files of the control samples}

\item{labels}{A label for every file, indicating to which batch it
belongs, e.g. the plate ID.}

\item{channels}{Names of the channels to normalize}

\item{transformList}{Transformation list to pass to the flowCore
\code{transform} function}

\item{nQ}{Number of quantiles to use. Default = 21, which results in
quantiles for every 5 percent of the data.}

\item{spar}{Smoothing parameter, typically between (0,1]. Default=0.5
See \code{smooth.spline} from the \code{stats} package}

\item{verbose}{If TRUE, progress updates are printed. Default = FALSE.}

\item{plot}{If TRUE, a plot is generated (using the \code{layout}
function) showing all quantiles. Default = FALSE.}

\item{plotTitle}{Title to use in the plot. Default = "Quantiles".}
}
\value{
A list containing all the splines and quantile information. This can
        be used as input for the \code{\link{QuantileNorm.normalize}} function.
}
\description{
Learn the batch effects from control samples.
This function computes quantiles to describe the distribution of the data,
and infers spline functions to equalize these distributions over the files.
Typically, you will use the function \code{\link{QuantileNorm.normalize}}
after this function to reverse the batch effects on other files.
}
\examples{

}
\seealso{
\code{\link{CytofNorm.train}}, \code{\link{QuantileNorm.normalize}}
}
