% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoNorm.R
\name{CytoNorm.train}
\alias{CytoNorm.train}
\title{CytoNorm.train}
\usage{
CytoNorm.train(
  files,
  labels,
  channels,
  transformList,
  outputDir = "./tmp",
  FlowSOM.params = list(nCells = 1e+06, xdim = 15, ydim = 15, nClus = 10, scale =
    FALSE),
  normMethod.train = QuantileNorm.train,
  normParams = list(nQ = 101),
  seed = NULL,
  clean = TRUE,
  plot = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{files}{Full paths to the fcs files or a flowSet of the control 
samples.}

\item{labels}{A label for every file, indicating to which batch it
belongs, e.g. the plate ID.}

\item{channels}{Column names of the channels that need to be normalized}

\item{transformList}{Transformation list to pass to the flowCore
\code{transform} function}

\item{outputDir}{Directory to put the temporary files in. Default = "./tmp"}

\item{FlowSOM.params}{Extra parameters to be passed to the FlowSOM
function, such as the number of cells
(\code{nCells}), the \code{xdim} and \code{y dim} or
the number of clusters (\code{nClus})}

\item{normMethod.train}{Normalization method to use for each cluster.
Default = \code{\link{QuantileNorm.train}}}

\item{normParams}{Parameters to pass to the normalization method. Default,
assuming \code{\link{QuantileNorm.train}}:
list(nQ = 21)). nQ is the number of quantiles
to use.}

\item{seed}{Set a seed for reproducable results.}

\item{clean}{Whether to remove the temporary files again at the end.
Default = TRUE.}

\item{plot}{If TRUE, plots are saved to the current dir.
Default = FALSE.}

\item{verbose}{If TRUE, progress updates are printed. Default = FALSE.}

\item{...}{Additional arguments to pass to read.FCS}
}
\value{
A list containing two elements: the FlowSOM clustering and a list
        containing all the splines per cluster. This can be used as input
        for the \code{\link{CytoNorm.normalize}} function.
}
\description{
CytoNorm learns batch effects, using control samples which should be the
the same per batch. The cells are first split into rough groups using the
FlowSOM algorithm, after which splines are trained to map the quantiles for
each control group to an average distribution
Typically, you will use the function \code{\link{CytoNorm.normalize}} after
this function to reverse the batch effects in other files.
}
\details{
Temporary fcs files splitting the data into clusters are written to
\code{outputDir} and will be removed again by default (depending on
\code{clean}).
}
\examples{

dir <- system.file("extdata", package = "CytoNorm")
files <- list.files(dir, pattern = "fcs$")
data <- data.frame(File = files,
                   Path = file.path(dir, files),
                   Type = stringr::str_match(files, "_([12]).fcs")[,2],
                   Batch = stringr::str_match(files, "PTLG[0-9]*")[,1],
                   stringsAsFactors = FALSE)
data$Type <- c("1" = "Train", "2" = "Validation")[data$Type]
train_data <- dplyr::filter(data, Type == "Train")
validation_data <- dplyr::filter(data, Type == "Validation")

ff <- flowCore::read.FCS(data$Path[1])
channels <- grep("Di$", flowCore::colnames(ff), value = TRUE)
transformList <- flowCore::transformList(channels,
                                         cytofTransform)
transformList.reverse <- flowCore::transformList(channels,
                                                 cytofTransform.reverse)

model <- CytoNorm.train(files = train_data$Path,
                        labels = train_data$Batch,
                        channels = channels,
                        transformList = transformList,
                        FlowSOM.params = list(nCells = 10000, #1000000
                                              xdim = 10,
                                              ydim = 10,
                                              nClus = 10,
                                              scale = FALSE),
                        normParams = list(nQ = 101),
                        seed = 1)

CytoNorm.normalize(model = model,
                   files = validation_data$Path,
                   labels = validation_data$Batch,
                   transformList = transformList,
                   transformList.reverse = transformList.reverse)

}
\seealso{
\code{\link{QuantileNorm.train}}, \code{\link{CytoNorm.normalize}},
           \code{\link{prepareFlowSOM}}
}
