% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoNorm.R
\name{prepareFlowSOM}
\alias{prepareFlowSOM}
\title{prepareFlowSOM}
\usage{
prepareFlowSOM(files, colsToUse, nCells = 1e+06,
  FlowSOM.params = list(xdim = 15, ydim = 15, nClus = 30, scale = FALSE),
  transformList = NULL, plot = FALSE, seed = NULL)
}
\arguments{
\item{files}{Files to use (with path)}

\item{colsToUse}{IDs or column names of the columns to use for clustering}

\item{nCells}{The total number of cells to use for the FlowSOM clustering.
This number is divided by the number of files to determine
the amount to select from each individual file.
Default = 1 000 000.}

\item{FlowSOM.params}{List with parameters to pass to the FlowSOM algorithm.
Default = \code{list(xdim = 15, ydim = 15, nclus = 30,
scale = FALSE)}.}

\item{transformList}{Transformation list to pass to the flowCore
\code{transform} function. Default = NULL.}

\item{plot}{If TRUE, the FlowSOM tree is plotted. Default = FALSE.}

\item{seed}{If not NULL, set.seed is called with this argument for
reproducable results. Default = NULL.}
}
\value{
FlowSOM object describing the FlowSOM clustering
}
\description{
Aggregate files, transform them and run FlowSOM.
This is used as a first step in the normalization procedure to detect
groups of similar cells. Typically you will not call this function, but use
the wrapper function \code{\link{CytoNorm.train}} instead.
}
\examples{

}
